+++ CHANGELOG.md
@@ -46,8 +46,10 @@
 - Update gitpod/workspace-base Docker digest to 0f38224 ([#3048](https://github.com/nf-core/tools/pull/3048))
 - update output_dir for api docs to new website structure ([#3051](https://github.com/nf-core/tools/pull/3051))
 - Update pre-commit hook astral-sh/ruff-pre-commit to v0.5.1 ([#3052](https://github.com/nf-core/tools/pull/3052))
+- Add `--limit-output` argument for modules/subworkflow update ([#3047](https://github.com/nf-core/tools/pull/3047))
 - update api docs to new structure ([#3054](https://github.com/nf-core/tools/pull/3054))
 - Update to pytest v8 and move it to dev dependencies ([#3058](https://github.com/nf-core/tools/pull/3058))
+- handle new jsonschema error type ([#3061](https://github.com/nf-core/tools/pull/3061))
 
 ## [v2.14.1 - Tantalum Toad - Patch](https://github.com/nf-core/tools/releases/tag/2.14.1) - [2024-05-09]